+++ src/react/components/autocompleter/__tests__/helpers/container.js
@@ -1,8 +1,0 @@
-let div = document.getElementById('test-container');
-if (!div) {
-	div = document.createElement('div');
-	div.id = 'test-container';
-	document.body.appendChild(div);
-}
-
-export default div;